Login is no longer automatic on password enter key
We have two buttons for login and resister. A user trying to register
using the fields provided will be automatically logged in, even though
that is not their intension, and that login will fail because they have
not yet registered.

While not a bug, it appears to be a bug to the user.